// src/lib/images.generated.ts
export const images = {
// ...
  "brand/logo": {
    "width": 2004,
    "height": 1468,
    "variants": [
      {
        "src": "/images/brand/logo-200.webp",
        "width": 200
      },
      {
        "src": "/images/brand/logo-2004.webp",
        "width": 2004
      }
    ]
  }
} as const satisfies Record<string, ImageEntry>;